+++ src/main/java/com/example/minicommerce/client/OauthClient.java
@@ -1,6 +1,6 @@
 package com.example.minicommerce.client;
 
-import com.example.minicommerce.global.GetOauthTokenResponseDto;
+import com.example.minicommerce.global.dto.GetOauthTokenResponseDto;
 import com.example.minicommerce.global.ProviderType;
 
 public interface OauthClient {
