+++ src/main/java/com/example/minicommerce/controller/OauthController.java
@@ -5,7 +5,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.example.minicommerce.global.GetOauthTokenResponseDto;
+import com.example.minicommerce.global.dto.GetOauthTokenResponseDto;
 import com.example.minicommerce.service.OauthService;
 
 @RestController
